requirements/production.txt
requests==2.7.0
dogstatsd-python==0.5.6
path.py==5.1
git+https://github.com/edx/codejail.git@2.2#egg=codejail==2.2
six
epicbox